Reformulações Técnicas

Criado Interface IFinanciamento para aplicação das regras de negócio nas classes de implementação FinanciamentoSAC e FinanciamentoPrice.

Reformulação na função calcular() no arquivo ./controller/TelaFinanciamentoController.java - retirada a regra de negócio do cálculo das amortizações.

Adição de funcionalidade

Adicionado a opção de cálculo de financiamento através do tipo de amortização Price

+++ src/br/com/samuel/calculadorajuroscompostos/model/JurosCompostos.java
@@ -3,13 +3,13 @@ package br.com.samuel.calculadorajuroscompostos.model;
 public class JurosCompostos {
     private double capitalInicial;
     private double taxa;
-    private int tempo;
+    private double tempo;
 
-    public int getTempo() {
+    public double getTempo() {
         return tempo;
     }
 
-    public void setTempo(int tempo) {
+    public void setTempo(double tempo) {
         this.tempo = tempo;
     }
